pyproject.toml
[project]
name = "research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boruta>=0.4.3",
    "catboost>=1.2.10",
    "iv-woe-filter>=0.1.1",
    "matplotlib>=3.10.7",
    "openai>=2.31.0",
    "pandas>=3.0.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "shap>=0.51.0",
]
